iot-data-to-opensearch/requirements.txt
kafka-python==2.2.11
pydantic==2.11.5
pydantic-settings==2.9.1
opensearch-py==3.0.0